requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1